Sorting by and new file checking system

+++ pdfdown-public.py
@@ -1,48 +1,54 @@
 from bs4 import BeautifulSoup, SoupStrainer
 import httplib2
-import pickle
 import os
 import requests
 import ssl
 from urllib.parse import urlparse
 import urllib.parse
 import getpass
+import re
 
 URLS = {"Parallel Programming" : "https://spcl.inf.ethz.ch/Teaching/2020-pp/",
         "Algorithmen und Wahrscheinlichkeit" : "https://www.cadmo.ethz.ch/education/lectures/FS20/AW/index.html",
         "Analysis I" : "https://metaphor.ethz.ch/x/2020/fs/401-0212-16L/",
-        "Analysis I/Battilana" : "https://battilana.uk/analysis-i-fs20/"}
+        "Analysis I/Battilana" : "https://battilana.uk/analysis-i-fs20/",
+        "Digitec/Labs" : "https://safari.ethz.ch/digitaltechnik/spring2020/doku.php?id=labs",
+        "Digitec/Lectures" : "https://safari.ethz.ch/digitaltechnik/spring2020/doku.php?id=schedule"}
 
-# URLS = ["https://spcl.inf.ethz.ch/Teaching/2020-pp/",
-#         "https://www.cadmo.ethz.ch/education/lectures/FS20/AW/index.html",
-#         "https://metaphor.ethz.ch/x/2020/fs/401-0212-16L/"]
-
-usedLinks = []
+SORT_BY = {r"AW_T(\d)+" : "Serie", "Minitest" : "Ministests", "Lecture" : "Lectures", "solution" : "Solutions",
+            "Loesung" : "Solutions", "Note" : "Lecture Notes", "Serie" : "Serie",
+            r"PP-L(\d)+" : "Lectures", "exercise" : "Exercises", "assignment" : "Exercises"}
 
 auth = None
 
 def download(links):
     for dir in links:
         for link in links[dir] :
-            response = requests.get(link, allow_redirects=True, auth=auth)
-            #print("Url: {}, Respnse:{}".format(url, response))
-            if(response.status_code == 200):
-                filename = dir + "/" + link[link.rfind("/")+1:].replace("%20", " ")
-                os.makedirs(os.path.dirname(filename), exist_ok=True)
-                with open(filename, 'wb') as fd:
-                    fd.write(response.content)
-                    # add file to the list of downloaded objects
-                    usedLinks.append(link)
-                    fd.close
-                    print(filename)
+            head = requests.head(link, allow_redirects=True, auth=auth)
+            if "Content-Disposition" in head.headers.keys():
+                filename = re.findall("filename=(.+)", head.headers["Content-Disposition"])[0].replace("\"", "").replace(";", "")
             else:
-                print("Url: {}, Response: {}".format(link, response))
+                filename = link[link.rfind("/")+1:].replace("%20", " ")
+            filename = dir + "/" + sortBy(filename) + filename
+
+            if not os.path.isfile(filename):
+                response = requests.get(link, allow_redirects=True, auth=auth)
+
+                if(response.status_code == 200):
+                    os.makedirs(os.path.dirname(filename), exist_ok=True)
+                    with open(filename, 'wb') as fd:
+                        fd.write(response.content)
+
+                        # add file to the list of downloaded objects
+                        fd.close
+                        print(filename)
+                else:
+                    print("Url: {}, Response: {}".format(link, response))
 
 
 def checkLink(link):
     if (link[-4:] == ".pdf") or (link[-4:] == ".zip"):
-        if not (link in usedLinks):
-            return True
+        return True
 
 def getLinks(sites):
     links = {}
@@ -65,26 +71,17 @@ def getLinks(sites):
 
     return links
 
-try:
-    dmp = open(__file__ + ".dmp", "rb")
-    usedLinks = pickle.load(dmp)
-except EOFError:
-    dmp.close()
-    pass
-except FileNotFoundError:
-    pass
-else:
-    dmp.close()
-
+def sortBy(filename):
+    for match in SORT_BY.items():
+        if re.match(match[0], filename):
+            return match[1] + "/"
+    return ""
 
 userName = input("Username: ")
 if (userName != ""):
     userPassword = getpass.getpass()
     auth = (userName, userPassword)
 
+print("Starting...")
 links = getLinks(URLS)
 download(links)
-
-dmp = open(__file__ + ".dmp", "wb")
-pickle.dump(usedLinks, dmp)
-dmp.close()
